% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePriorityInfo.R
\docType{methods}
\name{ComputePriorityInfo}
\alias{ComputePriorityInfo}
\alias{ComputePriorityInfo,UnitPrioritization,contObsPredModelParam,ErrorMoments,DD-method}
\title{Return the complementary info abouth the prioritization of units.}
\usage{
ComputePriorityInfo(object, ObsPredModelParam, Moments, DD)


  \S4method{ComputePriorityInfo}{UnitPrioritization,contObsPredModelParam,ErrorMoments,DD}(object,
  ObsPredModelParam, Moments, DD)
}
\arguments{
\item{object}{Object of class \linkS4class{UnitPrioritization}.}

\item{ObsPredModelParam}{Object of class \linkS4class{contObsPredModelParam}.}

\item{Moments}{Object of class \linkS4class{ErrorMoments}.}

\item{DD}{Object of class DD.}
}
\value{
Returns a \linkS4class{data.table} with the complementary info.
}
\description{
\code{ComputePriorityInfo} returns a \linkS4class{data.table} with complementary
information about the prioritization of units in the optimization approach to selective editing.
It basically returns the error estimates, the prediction error standard deviates, the quantile of
the error moments per domain, the quantile of sampling weights per domain and the quantile of the
unit score per quantile.
}
\examples{
\dontrun{

UnitPriorParam <- new(Class = 'UnitPrioritizationParam',
                      UnitScFunction = 'MinkUnitSc',
                      ScFunctionParam =  list(alpha = 1, Weights = 1))

PrioritizeUnits(ErrorMoment, UnitPriorParam)

}

}

