% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPriority.R
\docType{methods}
\name{getPriority}
\alias{getPriority}
\alias{getPriority,UnitPrioritization-method}
\title{Return the units with their editing priority}
\usage{
getPriority(object)

\S4method{getPriority}{UnitPrioritization}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{UnitPrioritization}.}
}
\value{
Numeric vector with the values of priority.
}
\description{
\code{getPriority} extracts the slots \code{UnitPriority} and \code{Units} to
return a \linkS4class{data.table} with each unit and its editing priority.
}
\examples{
\dontrun{
source('R:/USIE/Proyecto_DepSel_VarQual/datos/PreparacionStQsAnalisis.R', encoding = 'UTF-8')
fitPar <- new(Class = 'fitParam', edData = ff_2011_model.StQ, rawData = fd_2011_model.StQ,
              selection = FALSE,  formula = 'Ocupacion_35.__11.1.3._ ~ ActivEcono_35.__2.1.1._',
              selParam = list())
getEdData(fitPar)
}


}
