% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrioritizeUnits.R
\docType{methods}
\name{PrioritizeUnits}
\alias{PrioritizeUnits}
\alias{PrioritizeUnits,ErrorMoments,UnitPrioritizationParam-method}
\title{\code{PrioritizeUnits} computes the prioritization of units}
\usage{
PrioritizeUnits(object, Param)

\S4method{PrioritizeUnits}{ErrorMoments,UnitPrioritizationParam}(object, Param)
}
\arguments{
\item{object}{Object of class \linkS4class{SelEditErrorMoment}.}

\item{Param}{Object of class \linkS4class{UnitPrioritizationParam}.}
}
\value{
Object of class \linkS4class{UnitPrioritization}.
}
\description{
\code{PrioritizeUnits} computes the prioritization of units
}
\examples{
\dontrun{

UnitPriorParam <- new(Class = 'UnitPrioritizationParam',
                      UnitScFunction = 'MinkUnitSc',
                      ScFunctionParam =  list(alpha = 1, Weights = 1))

PrioritizeUnits(ErrorMoment, UnitPriorParam)

}
}
