% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitPrioritization-class.R
\docType{class}
\name{UnitPrioritization-class}
\alias{UnitPrioritization-class}
\title{S4 class for the prioritization of units}
\description{
Definition of the S4 class named \code{UnitPrioritization} for the unit
prioritization in the optimization approach to selective editing.
}
\section{Slots}{

\describe{
\item{\code{Domains}}{\linkS4class{data.table} with a row per domain with the values of the variables
determining the domains of the data set.}

\item{\code{Units}}{List of data.tables with as many components as domains (in the same order of the rows
of slot \code{Domains}) with every unit per domain.}

\item{\code{UnitScores}}{List with as many components as domains (in the same order of the rows
of slot \code{Domains}) with the scores of each unit (in the same order of the corresponding
data.table in slot \code{Units}).}

\item{\code{UnitPriority}}{List with as many components as domains (in the same order of the rows
of slot \code{Domains}) with the priority of each unit (in the same order of the corresponding
data.table in slot \code{Units}).}

\item{\code{PriorityInfo}}{Object of class \linkS4class{StQ} with complementary information about the
prioritization of units.}
}}

\examples{
# An empty UnitPrioritization object:
new(Class = 'UnitPrioritization')

\dontrun{
UnitPriorParam <- new(Class = 'UnitPrioritization')

}

}
