% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPriorityInfo.R
\docType{methods}
\name{setPriorityInfo<-}
\alias{setPriorityInfo<-}
\alias{setPriorityInfo<-,UnitPrioritization,StQ-method}
\title{Set complementary info about the prioritization of units.}
\usage{
setPriorityInfo(object) <- value

\S4method{setPriorityInfo}{UnitPrioritization,StQ}(object) <- value
}
\arguments{
\item{object}{Object of class \linkS4class{UnitPrioritization}.}

\item{value}{Object of class \linkS4class{StQ} with the info on the prioritization of units.}
}
\value{
Object of class \linkS4class{UnitPrioritization}.
}
\description{
This setter sets the slot PriorityInfo in a \linkS4class{UnitPrioritization} object.
}
