% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitPrioritization-class.R
\docType{class}
\name{UnitPrioritization-class}
\alias{UnitPrioritization-class}
\title{S4 class for the prioritization of units}
\description{
Definition of the S4 class named \code{UnitPrioritization} for the unit
prioritization in the optimization approach to selective editing.
}
\section{Slots}{

\describe{
\item{\code{Domains}}{\linkS4class{data.table} with the values of the variables indicating the domains.}

\item{\code{Units}}{list with the units of each domain.}

\item{\code{UnitScores}}{list with the scores of each unit.}

\item{\code{UnitPriority}}{list with the priority.}

\item{\code{PriorityInfo}}{list with complementary information about the prioritization of units.}
}}

\examples{
# An empty UnitPrioritization object:
new(Class = 'UnitPrioritization')

\dontrun{
UnitPriorParam <- new(Class = 'UnitPrioritization')

}

}
